Moved gitlab data objects to own package
