+++ src/main/java/com/dabsquared/gitlabjenkins/data/Branch.java
@@ -1,0 +1,44 @@
+package com.dabsquared.gitlabjenkins.data;
+
+public class Branch {
+
+    private String name;
+
+    private String ssh_url;
+
+    private String http_url;
+
+    private String namespace;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getSsh_url() {
+        return ssh_url;
+    }
+
+    public void setSsh_url(String ssh_url) {
+        this.ssh_url = ssh_url;
+    }
+
+    public String getHttp_url() {
+        return http_url;
+    }
+
+    public void setHttp_url(String http_url) {
+        this.http_url = http_url;
+    }
+
+    public String getNamespace() {
+        return namespace;
+    }
+
+    public void setNamespace(String namespace) {
+        this.namespace = namespace;
+    }
+}
