+++ src/main/java/com/dabsquared/gitlabjenkins/data/ObjectAttributes.java
@@ -1,0 +1,212 @@
+package com.dabsquared.gitlabjenkins.data;
+
+import org.apache.commons.lang.builder.ToStringBuilder;
+import org.apache.commons.lang.builder.ToStringStyle;
+import org.gitlab.api.models.GitlabUser;
+
+import java.util.Date;
+
+public class ObjectAttributes {
+
+    private Integer id;
+
+    private Integer iid;
+
+    private String targetBranch;
+
+    private String sourceBranch;
+
+    private Integer sourceProjectId;
+
+    private User author;
+
+    private User assignee;
+
+    private String title;
+
+    private Date createdAt;
+
+    private Date updatedAt;
+
+    private String state;
+
+    private String mergeStatus;
+
+    private Integer targetProjectId;
+
+    private String description;
+
+    private Branch source;
+
+    private Branch target;
+
+    private LastCommit lastCommit;
+
+    private String action;
+
+    public ObjectAttributes() {
+    }
+
+    @Override
+    public String toString() {
+        return ToStringBuilder.reflectionToString(this, ToStringStyle.MULTI_LINE_STYLE);
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public Integer getIid() {
+        return iid;
+    }
+
+    public void setIid(Integer iid) {
+        this.iid = iid;
+    }
+
+    public String getTargetBranch() {
+        return targetBranch;
+    }
+
+    public void setTargetBranch(String targetBranch) {
+        this.targetBranch = targetBranch;
+    }
+
+    public String getSourceBranch() {
+        return sourceBranch;
+    }
+
+    public void setSourceBranch(String sourceBranch) {
+        this.sourceBranch = sourceBranch;
+    }
+
+    public Integer getSourceProjectId() {
+        return sourceProjectId;
+    }
+
+    public void setSourceProjectId(Integer sourceProjectId) {
+        this.sourceProjectId = sourceProjectId;
+    }
+
+    public User getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(User author) {
+        this.author = author;
+    }
+
+    public void setAuthor(GitlabUser author) {
+        this.author = new User();
+        this.author.setId(author.getId());
+        this.author.setEmail(author.getEmail());
+        this.author.setName(author.getName());
+    }
+
+    public User getAssignee() {
+        return assignee;
+    }
+
+    public void setAssignee(User assignee) {
+        this.assignee = assignee;
+    }
+
+    public void setAssignee(GitlabUser assignee) {
+        this.assignee = new User();
+        this.assignee.setId(assignee.getId());
+        this.assignee.setEmail(assignee.getEmail());
+        this.assignee.setName(assignee.getName());
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public Date getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(Date createdAt) {
+        this.createdAt = createdAt;
+    }
+
+    public Date getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void setUpdatedAt(Date updatedAt) {
+        this.updatedAt = updatedAt;
+    }
+
+    public String getState() {
+        return state;
+    }
+
+    public void setState(String state) {
+        this.state = state;
+    }
+
+    public String getMergeStatus() {
+        return mergeStatus;
+    }
+
+    public void setMergeStatus(String mergeStatus) {
+        this.mergeStatus = mergeStatus;
+    }
+
+    public Integer getTargetProjectId() {
+        return targetProjectId;
+    }
+
+    public void setTargetProjectId(Integer targetProjectId) {
+        this.targetProjectId = targetProjectId;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Branch getSource() {
+        return source;
+    }
+
+    public void setSource(Branch source) {
+        this.source = source;
+    }
+
+    public Branch getTarget() {
+        return target;
+    }
+
+    public void setTarget(Branch target) {
+        this.target = target;
+    }
+
+    public LastCommit getLastCommit() {
+        return lastCommit;
+    }
+
+    public void setLastCommit(LastCommit lastCommit) {
+        this.lastCommit = lastCommit;
+    }
+
+    public String getAction() {
+        return action;
+    }
+
+    public void setAction(String action) {
+        this.action = action;
+    }
+}
